backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pandas>=2.2.0
requests>=2.31.0
python-multipart>=0.0.6
google-generativeai
python-dotenv
scikit-learn
gspread
